Damage & Stats!
Added Health node, that will be used in enemy and player's base nodes.
Added Stats resource. Will be used on every game object that requires such. Effects now change variables in those resources.
Damage and healing are functional. In its fullest, it seems. Armor and heal-res are yet to be polished.
To the towers I go!